backend/requirements/base.txt
Django==3.0.7
djangorestframework==3.11.0
drf-extensions==0.6.0
psycopg2-binary==2.8.5
gunicorn==20.0.4
